more tests

+++ src/main/java/br/com/helpdev/workshopunittest/controller/dto/FinanceSimulateResponseDTO.java
@@ -2,6 +2,6 @@ package br.com.helpdev.workshopunittest.controller.dto;
 
 public class FinanceSimulateResponseDTO {
 
-  public Long parcelValue;
+  public float parcelValue;
 
 }
